+++ src/main/java/kimsy/groceryapi/exception/presentation/WebClientExceptionController.java
@@ -1,6 +1,7 @@
 package kimsy.groceryapi.exception.presentation;
 
 import kimsy.groceryapi.exception.presentation.dto.ExceptionResult;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 
+@Slf4j
 @ControllerAdvice
 public class WebClientExceptionController {
     @ResponseStatus(HttpStatus.NOT_FOUND)
@@ -22,11 +24,20 @@ public class WebClientExceptionController {
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(WebClientResponseException.BadRequest.class)
-    public String badRequestHandler(Model model) {
+    public String badRequestHandler(WebClientResponseException e, Model model) {
+        log.info("외부 API에서 예외 발생 : {}", e.getResponseBodyAs(String.class));
+
         final ExceptionResult exceptionResult = new ExceptionResult(HttpStatus.BAD_REQUEST.name(),
                 HttpStatus.BAD_REQUEST.value(), "잘못된 요청입니다.");
 
         model.addAttribute("exceptionResult", exceptionResult);
         return "/error/4xx";
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(WebClientResponseException.InternalServerError.class)
+    public String internalServerErrorHandler(WebClientResponseException e , Model model) {
+        e.getMessage();
+        return "";
     }
 }
