+++ src/test/java/kimsy/groceryapi/product/application/ProductServiceTest.java
@@ -10,9 +10,9 @@ import java.util.Map;
 import kimsy.groceryapi.product.application.dto.ProductPriceResponse;
 import kimsy.groceryapi.product.application.dto.ProductsResponse;
 import kimsy.groceryapi.product.domain.AccessToken;
-import kimsy.groceryapi.product.domain.FruitWebClient;
+import kimsy.groceryapi.product.domain.web_client.FruitWebClient;
 import kimsy.groceryapi.product.domain.ProductWebClientAdapter;
-import kimsy.groceryapi.product.domain.VegetableWebClient;
+import kimsy.groceryapi.product.domain.web_client.VegetableWebClient;
 import kimsy.groceryapi.product.presentation.dto.ProductPriceRequest;
 import kimsy.groceryapi.product.presentation.dto.ProductsSearchRequest;
 import okhttp3.mockwebserver.MockResponse;
